EE175 - Senior Design/BMS_Dashboard/requirements-dev.txt
-r requirements.txt
Pillow>=10.3
pyinstaller>=6.8
